+++ src/Bridge/DoctrineStatsBundle/DependencyInjection/Configuration.php
@@ -1,0 +1,26 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Steevanb\DoctrineStats\Bridge\DoctrineStatsBundle\DependencyInjection;
+
+use Symfony\Component\Config\{
+    Definition\Builder\TreeBuilder,
+    Definition\ConfigurationInterface
+};
+
+class Configuration implements ConfigurationInterface
+{
+    public function getConfigTreeBuilder(): TreeBuilder
+    {
+        $treeBuilder = new TreeBuilder('doctrine_stats');
+        $rootNode = $treeBuilder->getRootNode();
+
+        $rootNode
+            ->children()
+                ->booleanNode('query_backtrace')->defaultValue(false)->end()
+            ->end();
+
+        return $treeBuilder;
+    }
+}