+++ src/Bridge/DoctrineStatsBundle/DoctrineStatsBundle.php
@@ -1,0 +1,22 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Steevanb\DoctrineStats\Bridge\DoctrineStatsBundle;
+
+use Steevanb\DoctrineStats\Bridge\DoctrineStatsBundle\DependencyInjection\Compiler\AddSqlLoggerPass;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class DoctrineStatsBundle extends Bundle
+{
+    /**
+     * @param ContainerBuilder $container
+     */
+    public function build(ContainerBuilder $container)
+    {
+        parent::build($container);
+
+        $container->addCompilerPass(new AddSqlLoggerPass());
+    }
+}